requirements.txt
Django==5.2.4
mysqlclient==2.2.7
pillow==11.3.0
django-auditlog==3.2.1
whitenoise==6.9.0
requests==2.32.4
openpyxl==3.1.5
pandas==2.3.1
django-extensions==4.1
djangorestframework==3.16.1 
django-cors-headers==4.7.0
gunicorn==23.0.0
drf-yasg==1.21.10
python-decouple==3.8
djangorestframework_simplejwt==5.5.0
dnspython==2.8.0 


